code refactor and navbar sticky and resolve comments
