+++ src/app/app.component.html
@@ -1,6 +1,6 @@
 <!--<h1 class="display-1 text-center py-3">Quize App</h1>-->
 <!--navbar start-->
-<nav class="navbar navbar-expand-lg bg-light">
+<nav class="navbar sticky-top navbar-expand-lg bg-light">
   <div class="container">
     <a class="navbar-brand" href="#">
       <img src="https://getbootstrap.com/docs/5.2/assets/brand/bootstrap-logo.svg" alt="Bootstrap" width="30" height="24">
