--- src/app/pages/product-listing/product-listing.component.css
@@ -1,11 +1,0 @@
-.p-outer {
-  display: flex;
-
-
-}
-.p-container {
-  display: flex;
-  flex-basis: 30%;
-  gap: 20px;
-
-}
