+++ src/app/shared/constants/api-end-point.ts
@@ -1,0 +1,1 @@
+export const PRODUCT_API = "https://fakestoreapi.com/products";
