+++ src/app/shared/product.model.ts
@@ -1,0 +1,14 @@
+export interface IProduct {
+  id: number
+  title: string
+  price: number
+  description: string
+  category: string
+  image: string
+  rating: IRating
+}
+
+export interface IRating {
+  rate: number
+  count: number
+}
